added IAdminDatabaseInterface and changed names
