+++ backend/src/main/java/se/cholmers/backend/Model/Order.java
@@ -7,13 +7,13 @@ import java.util.ArrayList;
 import java.util.List;
 
 import se.cholmers.backend.RequestException;
-import se.cholmers.backend.newDatabaseInterface;
+import se.cholmers.backend.DatabaseInterface;
 import se.cholmers.backend.Interface.IDatabaseInterface;
 
 public class Order {
     private LocalDateTime timeStamp;
     private List<IProduct> products;
-    private IDatabaseInterface dbi = newDatabaseInterface.getInstance();
+    private IDatabaseInterface dbi = DatabaseInterface.getInstance();
 
     //for testing
     public Order(List<IProduct> products){
