+++ backend/src/main/java/se/cholmers/backend/Model/Product.java
@@ -6,12 +6,11 @@ import org.yaml.snakeyaml.events.Event.ID;
 
 import se.cholmers.backend.DatabaseInterface;
 import se.cholmers.backend.RequestException;
-import se.cholmers.backend.newDatabaseInterface;
 import se.cholmers.backend.Interface.IDatabaseInterface;
 import se.cholmers.backend.Model.Interfaces.IProduct;
 
 public class Product implements IProduct {
-    private IDatabaseInterface dbi = newDatabaseInterface.getInstance();
+    private IDatabaseInterface dbi = DatabaseInterface.getInstance();
     private Integer amount;
     private String name;
     private String productID;
