+++ backend/src/main/java/se/cholmers/backend/Model/UserGroup.java
@@ -9,11 +9,11 @@ import se.cholmers.backend.Model.Interfaces.IOrderHistory;
 import se.cholmers.backend.Model.Interfaces.IProduct;
 import se.cholmers.backend.Model.Interfaces.IUserGroup;
 import se.cholmers.backend.RequestException;
-import se.cholmers.backend.newDatabaseInterface;
+import se.cholmers.backend.DatabaseInterface;
 import se.cholmers.backend.Interface.IDatabaseInterface;
 
 class UserGroup implements IUserGroup{
-    private IDatabaseInterface dbi = newDatabaseInterface.getInstance();
+    private IDatabaseInterface dbi = DatabaseInterface.getInstance();
     private Set<IProduct> products = new HashSet<>();
     private Year year;
     private String name;
